cleanup types and files
