+++ packages/api-core/src/index.ts
@@ -76,7 +76,6 @@ export * from './site-migration-status';
 export * from './site-owner-transfer';
 export * from './site-plans';
 export * from './site-plugins';
-export * from './site-update-schedules';
 export * from './site-preview-links';
 export * from './site-profiler';
 export * from './site-purchases';
@@ -87,6 +86,6 @@ export * from './site-staging-site';
 export * from './site-staging-sync';
 export * from './site-stats';
 export * from './site-themes';
-export * from './site-plugins';
+export * from './site-update-schedules';
 export * from './site-users';
 export * from './upgrades';
